+++ packages/fuselage/src/components/TextAreaInput/index.ts
@@ -1,0 +1,1 @@
+export * from './TextAreaInput';
