--- packages/fuselage/src/components/TextAreaInput/index.tsx
@@ -1,1 +1,0 @@
-export * from './TextArea';
